3/27/15 Major code revision. Script name changes, functionality changes in encode_batch_upload to use folder for insteadof direct input from mata_data_to_json.py
Functional changes for this patch reflect the usage of a small fastq upload to dev enviroment and  test log file writing Must be updated for usage in production and urls changed to reflect final upload.

Previous versions work but are to be used for reference and backup; they do not reflet current functionality of this patch.